chore(asgardian): separate to multiple files + cleanup
